Added highscore and made it possible to restart game
Added a highscore variable. Code now checks if current score is higher than highscore, if it is then highscore will be set to the current score. On death, current score and highscore will show on screen.

Made it possible to start a new round when game is over. Press space bar to start new game.